Modify multiple files and added a new file for login feature purposes

+++ SpringCleanProject/SpringClean/SpringClean/src/main/java/com/project/SpringClean/model/Customer.java
@@ -15,7 +15,7 @@ public class Customer {
     private String firstName;
     private String lastName;
     private String email;
-    private int phoneNumber;
+    private Long phoneNumber;
     private String address;
     private String password;
